+++ RELEASES.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## [Unreleased]
+## [v0.9.1] - 2023-11-19
 
 ### Changed
 
@@ -275,7 +275,7 @@
 * Initial public release
 
 
-[Unreleased]: https://github.com/powercap/raplcap/compare/v0.9.0...HEAD
+[v0.9.1]: https://github.com/powercap/raplcap/compare/v0.9.0...v0.9.1
 [v0.9.0]: https://github.com/powercap/raplcap/compare/v0.8.0...v0.9.0
 [v0.8.0]: https://github.com/powercap/raplcap/compare/v0.7.0...v0.8.0
 [v0.7.0]: https://github.com/powercap/raplcap/compare/v0.6.0...v0.7.0